+++ controlplane/src/core/services/AdmissionWebhookController.ts
@@ -5,9 +5,12 @@ import axiosRetry, { exponentialDelay } from 'axios-retry';
 import { FastifyBaseLogger } from 'fastify';
 import { OrganizationEventName } from '@wundergraph/cosmo-connect/dist/notifications/events_pb';
 import { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
+import { HttpProxyAgent } from 'http-proxy-agent';
+import { HttpsProxyAgent } from 'https-proxy-agent';
 import * as schema from '../../db/schema.js';
 import { WebhookDeliveryInfo } from '../../db/models.js';
 import { webhookAxiosRetryCond } from '../util.js';
+import { traced } from '../tracing.js';
 
 export class AdmissionError extends Error {
   constructor(message: string, cause?: Error) {
@@ -32,6 +35,7 @@ export interface ValidateConfigResponse {
   error?: string;
 }
 
+@traced
 export class AdmissionWebhookController {
   httpClient: AxiosInstance;
   constructor(
@@ -39,8 +43,23 @@ export class AdmissionWebhookController {
     private logger: FastifyBaseLogger,
     private graphAdmissionWebhookURL?: string,
     private graphAdmissionWebhookSecret?: string,
+    proxyUrl?: string,
   ) {
+    let httpAgent: HttpProxyAgent<string> | undefined;
+    let httpsAgent: HttpsProxyAgent<string> | undefined;
+    if (proxyUrl) {
+      try {
+        httpAgent = new HttpProxyAgent(proxyUrl, {});
+        httpsAgent = new HttpsProxyAgent(proxyUrl, {});
+      } catch (e) {
+        logger.error(e, 'Failed to create proxy agent');
+      }
+    }
+
     this.httpClient = axios.create({
+      httpAgent,
+      httpsAgent,
+      proxy: false,
       timeout: 30_000,
       baseURL: this.graphAdmissionWebhookURL,
     });