+++ router-tests/events/nats_events_test.go
@@ -56,7 +56,6 @@ func (c *ConfigPollerMock) Stop(_ context.Context) error {
 	return nil
 }
 
-
 type natsSubscriptionArgs struct {
 	dataValue []byte
 	errValue  error
@@ -1431,7 +1430,6 @@ func TestNatsEvents(t *testing.T) {
 			})
 		})
 	})
-
 
 	t.Run("subscribe to multiple subjects", func(t *testing.T) {
 		t.Parallel()